requirements.txt
aiosmtpd==1.4.6
alembic==1.13.3
atpublic==5.0
attrs==24.2.0
blinker==1.8.2
click==8.1.7
distlib==0.3.8
dnspython==2.6.1
email_validator==2.2.0
filelock==3.16.1
Flask==3.0.3
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-Moment==1.0.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
platformdirs==4.3.6
PyJWT==2.9.0
python-dotenv==1.0.1
SQLAlchemy==2.0.35
typing_extensions==4.12.2
virtualenv==20.26.5
Werkzeug==3.0.4
WTForms==3.1.2
